pyproject.toml
[tool.poetry]
name = "apiswymmer"
version = "0.1.0"
description = "Swagger Doc Generator"
authors = ["Anuran Roy <anuranroy02@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Jinja2 = "3.1.2"
markdown-it-py = "3.0.0"
MarkupSafe = "2.1.3"
mdurl = "0.1.2"
Pygments = "2.16.1"
PyYAML = "6.0.1"
rich = "13.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
